refactor(membershipStats): split resolver into mult. files
